@@ -1,6 +1,10 @@
-module.exports = {
+/** @type {import('tailwindcss').Config} */
+export default {
 	darkMode: 'media',
-	content: ['./index.html', './src/**/*.{svelte,js,ts,html}'],
+	content: [
+		"./index.html",
+		"./src/**/*.{svelte,js,ts,jsx,tsx}",
+	],
 	theme: {
 		extend: {
 			spacing: {
@@ -28,4 +32,4 @@ module.exports = {
 			strategy: 'class'
 		})
 	]
-};
+}
