@@ -1,8 +1,9 @@
-import App from './App.svelte';
-import './index.css';
+import { mount } from 'svelte'
+import './app.css'
+import App from './App.svelte'
 
-const app = new App({
-	target: document.getElementById('app')
-});
+const app = mount(App, {
+  target: document.getElementById('app'),
+})
 
-export default app;
+export default app
