@@ -1,11 +1,7 @@
-import { defineConfig } from 'vite';
-import { svelte } from '@sveltejs/vite-plugin-svelte';
+import { defineConfig } from 'vite'
+import { svelte } from '@sveltejs/vite-plugin-svelte'
 
-// https://vitejs.dev/config/
+// https://vite.dev/config/
 export default defineConfig({
-	plugins: [svelte()],
-	build: {
-		outDir: 'S3',
-		emptyOutDir: true
-	}
-});
+  plugins: [svelte()],
+})
